refactor: question command-service 리팩토링

+++ be/post-command-service/src/main/java/com/zzuag/post_command_service/domain/question/service/QuestionService.java
@@ -3,8 +3,8 @@ package com.zzuag.post_command_service.domain.question.service;
 import com.zzuag.post_command_service.domain.question.dto.request.QuestionCreateRequest;
 import com.zzuag.post_command_service.domain.question.dto.request.QuestionEditRequest;
 import com.zzuag.post_command_service.domain.question.entity.QuestionEntity;
-import com.zzuag.post_command_service.domain.question.event.QuestionLikeEvent;
-import com.zzuag.post_command_service.domain.question.event.QuestionUnlikeEvent;
+import com.zzuag.post_command_service.domain.question.service.event.QuestionLikeEvent;
+import com.zzuag.post_command_service.domain.question.service.event.QuestionUnlikeEvent;
 import com.zzuag.post_command_service.domain.question.persistence.repository.QuestionEntityRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,23 +20,35 @@ public class QuestionService {
     private final QuestionEntityRepository questionEntityRepository;
 
     @Transactional
-    public void createQuestion(QuestionCreateRequest request){
-        QuestionEntity save = questionEntityRepository.save(QuestionEntity.create(request));
+    public void createQuestion(Long userId, QuestionCreateRequest request){
+        QuestionEntity save = questionEntityRepository.save(QuestionEntity.create(userId, request));
         log.info("Question created: {}", save);
     }
 
     @Transactional
-    public void editQuestion(Long questionId, QuestionEditRequest request){
+    public void editQuestion(Long questionId, Long userId, QuestionEditRequest request){
         QuestionEntity target = questionEntityRepository.findByIdAndIsDeletedFalse(questionId)
                 .orElseThrow(() -> new IllegalArgumentException("Question not found"));
+        if (target.isDeleted()){
+            throw new IllegalArgumentException("Question is deleted");
+        }
+        if(!target.getUserId().equals(userId)){
+            throw new IllegalArgumentException("User not authorized");
+        }
         target.update(request);
         log.info("Question edited: {}", target);
     }
 
     @Transactional
-    public void deleteQuestion(Long questionId){
+    public void deleteQuestion(Long userId, Long questionId){
         QuestionEntity target = questionEntityRepository.findByIdAndIsDeletedFalse(questionId)
                 .orElseThrow(() -> new IllegalArgumentException("Question not found"));
+        if (target.isDeleted()){
+            throw new IllegalArgumentException("Question is deleted");
+        }
+        if(!target.getUserId().equals(userId)){
+            throw new IllegalArgumentException("User not authorized");
+        }
         target.delete();
         log.info("Question deleted: {}", target);
     }
